Refactor in Process struct
Refactored processes data read in from file to be included in a process
struct. This eliminates having multiple vectors for each set of data and
allows reuse of the struct in the different schedulers. Refactored FCFS
method to use the process struct and removed the old method. Output should
be the same.